Seperated piServer.py into its server and hardware components
piServer.py is now server.py and hardware.py so that demonstration on desktop is easier to follow.
server.pi needs templates folder and static folder to work. index.html has been updated now to accomodate the server, however no longer works by itself and relies on the server.

hardware code is directly copied from the raspberry pi pi1 2011, and has been fully tested and debugged on that system. as it uses the GPIO library, it will not function on a desktop, and as such was why piServer needed to be seperated so that we can demonstrate the server on desktop.